[ESPACE] update code for version v2.0.0
