+++ .gitignore
@@ -7,11 +7,12 @@
 
 # Test binary, built with `go test -c`
 *.test
-test
 # Output of the go coverage tool, specifically when used with LiteIDE
 *.out
 
+app
+
 # Dependency directories (remove the comment below to include it)
 # vendor/
-
+.idea/
 go.sum
