+++ api/constant/constant.go
@@ -1,0 +1,13 @@
+package constant
+
+const  (
+	TABLE_INFO = 0
+	TABLE_PURE = 1
+	ERROR_TABLE_NOT_GENERATE = "Error! Table is not generated."
+)
+
+const (
+	SIZE_DEFAULT = 0
+	MAX_LENGTH = 9
+	FLOAT_ERROR = -0.000000000000000000000000000000000000001
+)
