+++ api/space/SpaceLoad.go
@@ -1,0 +1,4 @@
+package space
+
+type SpaceLoad interface {
+}
