+++ go.mod
@@ -2,4 +2,9 @@ module github.com/emtabb/espace
 
 go 1.14
 
-require github.com/emtabb/field v1.0.0
+require (
+	github.com/emtabb/field v1.0.0
+	github.com/emtabb/qugo v1.0.1
+	github.com/emtabb/state v1.0.1
+	go.mongodb.org/mongo-driver v1.5.0
+)
