--- src/space/util/arrays.go
@@ -1,9 +1,0 @@
-package util
-
-// func RemoveStrFromArray(data string, list []string) []string {
-
-// }
-
-// func RemoveFromArray(data interface {}, list []interface {}) []interface {} {
-
-// }
