+++ surface/init.go
@@ -1,6 +1,6 @@
 package surface
 
-import "github.com/emtabb/espace/src/space"
+import "github.com/emtabb/espace/api/space"
 func Surface() *space.SpaceStructure{
 	return new(space.SpaceStructure)
 }
